Moved TextRenderer into a separate file
