+++ packages/flame/lib/src/components/fps_text_component.dart
@@ -1,4 +1,5 @@
 import '../../components.dart';
+import '../text/text_renderer.dart';
 
 /// The [FpsTextComponent] is a [TextComponent] that writes out the current FPS.
 /// It has a [FpsComponent] as a child which does the actual calculations.
