+++ packages/flame/lib/text.dart
@@ -1,0 +1,2 @@
+
+export 'src/text/text_renderer.dart' show TextRenderer;
