+++ packages/flame/test/text_test.dart
@@ -1,4 +1,5 @@
 import 'package:flame/components.dart';
+import 'package:flame/src/text/text_renderer.dart';
 import 'package:flutter/material.dart';
 import 'package:test/test.dart';
